# Feature/improve popup trigger performance