--- a/frontend/src/app/pages/exercises/exercise/shared/exercise-map/utility/popup-helper.ts
+++ b/frontend/src/app/pages/exercises/exercise/shared/exercise-map/utility/popup-helper.ts
@@ -106,7 +106,6 @@ export class ImagePopupHelper {
     }
 
     /**
-     * @param layer The layer the feature is in
      * @returns wether the feature is visible at the given pixel
      */
     private hasFeatureAtPixel(feature: Feature<Point>, pixel: Pixel): boolean {