… backend/package.json
… shared/package.json