* Updates all packages
    * Only version bumps on minor and patch level
    * OpenLayers now sets `willReadFrequently` on the canvas rendering contexts (https://github.com/openlayers/openlayers/releases/tag/v7.2.0). A quick test showed this improves the performance from ca. 150ms per click to ca. 120ms. There was no difference when the map was not moved in between clicks.

* Improvements in `hasFeatureAtPixel` in `PopupManager`
    * Early return in `hasFeatureAtPixel`
    * Skip not relevant layers
    * This improved the performance from ca. 120 ms to ca. 35ms

While the used workaround is still very ugly, it is now at least less expensive and reduces the stutter when opening a popup.

The creation and rendering of the HTML popup were not included in the performance measurements above. They are only the time it took the JS from the registration of the click to the trigger of the rendering in the DOM.