@Dassderdie 2022-12-22
@ClFeSc FYI, I updated OpenLayers once again from `7.2.0` to `7.2.2`. The only things that changed were in some JSdoc custom scripts. I believe to fix the commonjs build. I couldn't notice any difference but wanted to make sure that we don't run into some weird problem later.